backend/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
groq>=0.11.0
google-generativeai>=0.8.0
pypdf>=5.0.0
python-docx>=1.1.0
supabase>=2.10.0
PyJWT[crypto]>=2.9.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
